+++ src/app.rs
@@ -1402,6 +1402,9 @@ fn flag_quiet(args: &mut Vec<RGArg>) {
 Do not print anything to stdout. If a match is found in a file, then ripgrep
 will stop searching. This is useful when ripgrep is used only for its exit
 code (which will be an error if no matches are found).
+
+When --files is used, then ripgrep will stop finding files after finding the
+first file that matches all ignore rules.
 ");
     let arg = RGArg::switch("quiet").short("q")
         .help(SHORT).long_help(LONG);