+++ CHANGELOG.md
@@ -84,6 +84,9 @@ Bug fixes:
   Upgrade `grep` crate to `regex-syntax 0.5.0`.
 * [BUG #893](https://github.com/BurntSushi/ripgrep/issues/893):
   Improve support for git submodules.
+* [BUG #907](https://github.com/BurntSushi/ripgrep/issues/907):
+  ripgrep will now stop traversing after the first file when `--quiet --files`
+  is used.
 * [BUG #918](https://github.com/BurntSushi/ripgrep/issues/918):
   Don't skip tar archives when `-z/--search-zip` is used.
 * [BUG #934](https://github.com/BurntSushi/ripgrep/issues/934):